refactor: extract reusable order hooks and components
