+++ app/dashboard/page.tsx
@@ -6,34 +6,11 @@ import { useRouter } from 'next/navigation';
 import { supabase, Order } from '../../lib/supabase';
 import Header from '../../components/Header';
 import TabBar from '../../components/TabBar';
-import OrderCard from './OrderCard';
 import CalendarView from './CalendarView';
 import UserManagement from './UserManagement';
-
-type QuoteStatus = 'pending' | 'responded' | 'accepted' | 'rejected';
-
-interface Quote {
-  id: string;
-  customer_name: string;
-  customer_email?: string | null;
-  customer_phone?: string | null;
-  occasion?: string | null;
-  age_group?: string | null;
-  theme?: string | null;
-  servings?: string | null;
-  budget?: string | null;
-  event_date?: string | null;
-  event_details?: string | null;
-  has_reference_photo?: boolean;
-  photo_description?: string | null;
-  reference_photo_url?: string | null;
-  status: QuoteStatus;
-  created_at: string;
-  estimated_price?: number | null;
-  admin_notes?: string | null;
-  responded_at?: string | null;
-  updated_at?: string | null;
-}
+import OrderList from './OrderList';
+import QuoteList from './QuoteList';
+import { Quote, QuoteStatus } from './types';
 
 const quoteStatusColors: Record<QuoteStatus, string> = {
   pending: 'bg-yellow-100 text-yellow-800',
@@ -54,10 +31,6 @@ export default function DashboardPage() {
   const [loading, setLoading] = useState(true);
   const [currentUser, setCurrentUser] = useState<any>(null);
   const [showSuccessMessage, setShowSuccessMessage] = useState('');
-  const [deleteConfirmation, setDeleteConfirmation] = useState<string | null>(null);
-  const [deletingQuote, setDeletingQuote] = useState<string | null>(null);
-  const [deleteOrderConfirmation, setDeleteOrderConfirmation] = useState<string | null>(null);
-  const [deletingOrder, setDeletingOrder] = useState<string | null>(null);
   const [employeeCancelRequest, setEmployeeCancelRequest] = useState<string | null>(null);
   const [pendingCancelRequests, setPendingCancelRequests] = useState<{[key: string]: any}>({});
   const router = useRouter();
@@ -424,8 +397,6 @@ export default function DashboardPage() {
   // Quote deletion handling
   // -------------------------------------------------------------------------
   const handleDeleteQuote = async (quoteId: string) => {
-    setDeletingQuote(quoteId);
-
     try {
       const { error } = await supabase
         .from('quotes')
@@ -448,17 +419,12 @@ export default function DashboardPage() {
       console.error('Error deleting quote:', error);
       alert('Error al eliminar la cotización. Por favor intenta de nuevo.');
     }
-
-    setDeletingQuote(null);
-    setDeleteConfirmation(null);
   };
 
   // -------------------------------------------------------------------------
   // Order deletion handling
   // -------------------------------------------------------------------------
   const handleDeleteOrder = async (orderId: string) => {
-    setDeletingOrder(orderId);
-
     try {
       const { error } = await supabase
         .from('orders')
@@ -482,9 +448,6 @@ export default function DashboardPage() {
       // Fallback to local deletion
       deleteOrderFromLocal(orderId);
     }
-
-    setDeletingOrder(null);
-    setDeleteOrderConfirmation(null);
   };
 
   const deleteOrderFromLocal = (orderId: string) => {
@@ -722,329 +685,25 @@ export default function DashboardPage() {
             <div className="p-4">
               {/* Orders Tab */}
               {activeTab === 'orders' && (
-                <div className="space-y-4">
-                  {showSuccessMessage && (
-                    <div className="bg-green-50 border border-green-200 rounded-lg p-4 mb-4">
-                      <div className="flex items-center">
-                        <div className="w-5 h-5 flex items-center justify-center">
-                          <i className="ri-check-line text-green-600"></i>
-                        </div>
-                        <p className="ml-3 text-green-700 font-medium">{showSuccessMessage}</p>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Delete Order Confirmation Modal */}
-                  {deleteOrderConfirmation && (
-                    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-                      <div className="bg-white rounded-xl p-6 max-w-sm mx-4 shadow-2xl">
-                        <div className="text-center">
-                          <div className="w-16 h-16 flex items-center justify-center bg-red-100 rounded-full mx-auto mb-4">
-                            <i className="ri-delete-bin-line text-red-500 text-2xl"></i>
-                          </div>
-                          <h3 className="text-lg font-bold text-gray-800 mb-2">¿Eliminar Pedido?</h3>
-                          <p className="text-gray-600 text-sm mb-6">
-                            Esta acción no se puede deshacer. El pedido será eliminado permanentemente.
-                          </p>
-                          <div className="flex space-x-3">
-                            <button
-                              onClick={() => setDeleteOrderConfirmation(null)}
-                              className="flex-1 bg-gray-200 text-gray-800 py-3 px-4 rounded-lg font-medium hover:bg-gray-300 transition-colors"
-                            >
-                              Cancelar
-                            </button>
-                            <button
-                              onClick={() => handleDeleteOrder(deleteOrderConfirmation)}
-                              disabled={deletingOrder === deleteOrderConfirmation}
-                              className="flex-1 bg-red-600 text-white py-3 px-4 rounded-lg font-medium hover:bg-red-700 transition-colors disabled:opacity-50"
-                            >
-                              {deletingOrder === deleteOrderConfirmation ? (
-                                <div className="flex items-center justify-center">
-                                  <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
-                                  Eliminando...
-                                </div>
-                              ) : (
-                                'Eliminar'
-                              )}
-                            </button>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  <div className="bg-gradient-to-r from-pink-400 to-teal-400 rounded-xl p-4 text-white">
-                    <div className="flex items-center justify-between">
-                      <div>
-                        <h3 className="text-lg font-semibold mb-1">Vista Rápida</h3>
-                        <p className="text-sm opacity-90">
-                          Gestiona los pedidos de hoy eficientemente
-                        </p>
-                      </div>
-                      <button
-                        onClick={() => setShowTodayView(true)}
-                        className="bg-white/20 backdrop-blur-sm text-white px-4 py-2 rounded-lg font-medium !rounded-button hover:bg-white/30 transition-all transform hover:scale-105 shadow-lg"
-                      >
-                        <div className="flex items-center justify-between">
-                          <div>
-                            <p className="font-medium text-white">Hoy</p>
-                            <p className="text-sm text-white/90">{getTodayOrders().length} pedido{getTodayOrders().length !== 1 ? 's' : ''}</p>
-                          </div>
-                          <div className="flex space-x-1 ml-3">
-                            <div className="w-6 h-6 rounded-full bg-white/30 text-white text-xs flex items-center justify-center font-medium">
-                              {getTodayOrders().length}
-                            </div>
-                            <i className="ri-arrow-right-line text-lg ml-2"></i>
-                          </div>
-                        </div>
-                      </button>
-                    </div>
-                  </div>
-
-                  {/* Orders list or empty state */}
-                  {orders.length === 0 ? (
-                    <div className="text-center py-8">
-                      <div className="w-16 h-16 flex items-center justify-center bg-gray-100 rounded-full mx-auto mb-4">
-                        <i className="ri-shopping-bag-line text-gray-400 text-2xl"></i>
-                      </div>
-                      <h3 className="text-lg font-semibold text-gray-800 mb-2">No hay pedidos aún</h3>
-                      <p className="text-gray-600 text-sm">
-                        Los pedidos aparecerán aquí cuando los clientes los
-                        realicen
-                      </p>
-                    </div>
-                  ) : (
-                    orders.map((order) => (
-                      <div key={order.id} className="bg-white rounded-2xl shadow-lg overflow-hidden hover:shadow-xl transition-shadow">
-                        <div className="p-5">
-                          <div className="flex items-center justify-between mb-4">
-                            <div className="flex items-center space-x-4">
-                              <div className="w-14 h-14 bg-gradient-to-br from-pink-400 via-purple-400 to-teal-400 rounded-full flex items-center justify-center text-white font-bold text-lg shadow-md">
-                                {order.customer_name.charAt(0)}
-                              </div>
-                              <div>
-                                <h3 className="font-bold text-gray-800 text-lg">{order.customer_name}</h3>
-                                <div className="flex items-center space-x-2 text-sm text-gray-600">
-                                  <i className="ri-time-line"></i>
-                                  <span>{order.pickup_time || 'Hora no especificada'}</span>
-                                </div>
-                              </div>
-                            </div>
-                            <div className="flex items-center space-x-2">
-                              <div className="text-right">
-                                <p className="text-2xl font-bold text-pink-600">${order.total.toFixed(2)}</p>
-                                <span
-                                  className={`px-3 py-1 rounded-full text-xs font-bold ${statusColors[order.status]} shadow-sm`}
-                                >
-                                  <i
-                                    className={`${
-                                      order.status === 'pending'
-                                        ? 'ri-timer-line'
-                                        : order.status === 'baking'
-                                        ? 'ri-fire-line'
-                                        : order.status === 'decorating'
-                                        ? 'ri-brush-line'
-                                        : order.status === 'ready'
-                                        ? 'ri-check-line'
-                                        : order.status === 'completed'
-                                        ? 'ri-check-double-line'
-                                        : 'ri-close-line'
-                                    } mr-1`}
-                                  ></i>
-                                  {order.status === 'pending'
-                                    ? 'Pendiente'
-                                    : order.status === 'baking'
-                                    ? 'Horneando'
-                                    : order.status === 'decorating'
-                                    ? 'Decorando'
-                                    : order.status === 'ready'
-                                    ? 'Listo'
-                                    : order.status === 'completed'
-                                    ? 'Completado'
-                                    : 'Cancelado'}
-                                </span>
-                              </div>
-                              {/* Only show delete button for owners */}
-                              {currentUser?.role === 'owner' && (
-                                <button
-                                  onClick={() => setDeleteOrderConfirmation(order.id)}
-                                  className="w-8 h-8 flex items-center justify-center bg-red-100 text-red-600 rounded-full hover:bg-red-200 transition-colors"
-                                  title="Eliminar pedido"
-                                >
-                                  <i className="ri-delete-bin-line text-sm"></i>
-                                </button>
-                              )}
-                            </div>
-                          </div>
-
-                          <div className="mb-4">
-                            <h4 className="text-sm font-bold text-gray-700 mb-2 flex items-center">
-                              <i className="ri-shopping-bag-line mr-2"></i>
-                              Items del pedido:
-                            </h4>
-                            <div className="bg-gray-50 rounded-xl p-3 space-y-2">
-                              {order.items.map((item, idx) => (
-                                <div key={idx} className="flex justify-between items-center">
-                                  <div className="flex items-center space-x-2">
-                                    <div className="w-6 h-6 bg-pink-100 rounded-full flex items-center justify-center text-pink-600 font-bold text-xs">
-                                      {item.quantity}
-                                    </div>
-                                    <span className="text-sm font-medium text-gray-700">{item.name}</span>
-                                  </div>
-                                  <span className="text-sm font-bold text-gray-800">{item.price}</span>
-                                </div>
-                              ))}
-                            </div>
-                          </div>
-
-                          {order.special_requests && (
-                            <div className="mb-4 p-3 bg-gradient-to-r from-yellow-50 to-amber-50 rounded-xl border-l-4 border-yellow-400">
-                              <div className="flex items-start space-x-2">
-                                <i className="ri-lightbulb-line text-yellow-600 mt-0.5"></i>
-                                <div>
-                                  <p className="text-sm font-bold text-yellow-800">Solicitud Especial:</p>
-                                  <p className="text-sm text-yellow-700">{order.special_requests}</p>
-                                </div>
-                              </div>
-                            </div>
-                          )}
-
-                          <div className="flex space-x-2">
-                            {order.status === 'pending' && (
-                              <button
-                                onClick={() => updateOrderStatus(order.id, 'baking')}
-                                className="flex-1 bg-gradient-to-r from-orange-500 to-red-500 text-white py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-orange-600 hover:to-red-600 transition-all transform hover:scale-105 shadow-lg"
-                              >
-                                <i className="ri-fire-line mr-2"></i>
-                                Iniciar Horneado
-                              </button>
-                            )}
-                            {order.status === 'baking' && (
-                              <button
-                                onClick={() => updateOrderStatus(order.id, 'decorating')}
-                                className="flex-1 bg-gradient-to-r from-purple-500 to-pink-500 text-white py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-purple-600 hover:to-pink-600 transition-all transform hover:scale-105 shadow-lg"
-                              >
-                                <i className="ri-brush-line mr-2"></i>
-                                Iniciar Decoración
-                              </button>
-                            )}
-                            {order.status === 'decorating' && (
-                              <button
-                                onClick={() => updateOrderStatus(order.id, 'ready')}
-                                className="flex-1 bg-gradient-to-r from-green-500 to-emerald-500 text-white py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-green-600 hover:to-emerald-600 transition-all transform hover:scale-105 shadow-lg"
-                              >
-                                <i className="ri-check-line mr-2"></i>
-                                Marcar Listo
-                              </button>
-                            )}
-                            {order.status === 'ready' && (
-                              <button
-                                onClick={() => updateOrderStatus(order.id, 'completed')}
-                                className="flex-1 bg-gradient-to-r from-blue-500 to-indigo-500 text-white py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-blue-600 hover:to-indigo-600 transition-all transform hover:scale-105 shadow-lg"
-                              >
-                                <i className="ri-check-double-line mr-2"></i>
-                                Completar Pedido
-                              </button>
-                            )}
-                            
-                            {/* Cancel button - different behavior for owners vs employees */}
-                            {order.status !== 'completed' && order.status !== 'cancelled' && (
-                              <button
-                                onClick={() => updateOrderStatus(order.id, 'cancelled')}
-                                className="bg-gradient-to-r from-red-500 to-red-600 text-white py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-red-600 hover:to-red-700 transition-all transform hover:scale-105 shadow-lg"
-                                title={currentUser?.role === 'employee' ? 'Solicitar cancelación al propietario' : 'Cancelar pedido'}
-                              >
-                                <i className="ri-close-line mr-2"></i>
-                                {currentUser?.role === 'employee' ? 'Solicitar Cancelar' : 'Cancelar'}
-                              </button>
-                            )}
-                            
-                            <button
-                              onClick={() => window.open(`tel:${order.customer_phone}`, '_self')}
-                              className="bg-gradient-to-r from-gray-100 to-gray-200 text-gray-700 py-3 px-4 rounded-xl text-sm font-bold !rounded-button hover:from-gray-200 hover:to-gray-300 transition-all transform hover:scale-105 shadow-lg"
-                            >
-                              <i className="ri-phone-line text-lg"></i>
-                            </button>
-                          </div>
-                        </div>
-                      </div>
-                    ))
-                  )}
-                </div>
+                <OrderList
+                  orders={orders}
+                  currentUser={currentUser}
+                  showSuccessMessage={showSuccessMessage}
+                  onStatusUpdate={updateOrderStatus}
+                  onDelete={handleDeleteOrder}
+                  setShowTodayView={setShowTodayView}
+                />
               )}
 
               {/* Quotes Tab */}
               {activeTab === 'quotes' && (
-                <div className="space-y-4">
-                  {showSuccessMessage && (
-                    <div className="bg-green-50 border border-green-200 rounded-lg p-4 mb-4">
-                      <div className="flex items-center">
-                        <div className="w-5 h-5 flex items-center justify-center">
-                          <i className="ri-check-line text-green-600"></i>
-                        </div>
-                        <p className="ml-3 text-green-700 font-medium">{showSuccessMessage}</p>
-                      </div>
-                    </div>
-                  )}
-
-                  {/* Delete Confirmation Modal */}
-                  {deleteConfirmation && (
-                    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-                      <div className="bg-white rounded-xl p-6 max-w-sm mx-4 shadow-2xl">
-                        <div className="text-center">
-                          <div className="w-16 h-16 flex items-center justify-center bg-red-100 rounded-full mx-auto mb-4">
-                            <i className="ri-delete-bin-line text-red-500 text-2xl"></i>
-                          </div>
-                          <h3 className="text-lg font-bold text-gray-800 mb-2">¿Eliminar Cotización?</h3>
-                          <p className="text-gray-600 text-sm mb-6">
-                            Esta acción no se puede deshacer. La cotización será eliminada permanentemente.
-                          </p>
-                          <div className="flex space-x-3">
-                            <button
-                              onClick={() => setDeleteConfirmation(null)}
-                              className="flex-1 bg-gray-200 text-gray-800 py-3 px-4 rounded-lg font-medium hover:bg-gray-300 transition-colors"
-                            >
-                              Cancelar
-                            </button>
-                            <button
-                              onClick={() => handleDeleteQuote(deleteConfirmation)}
-                              disabled={deletingQuote === deleteConfirmation}
-                              className="flex-1 bg-red-600 text-white py-3 px-4 rounded-lg font-medium hover:bg-red-700 transition-colors disabled:opacity-50"
-                            >
-                              {deletingQuote === deleteConfirmation ? (
-                                <div className="flex items-center justify-center">
-                                  <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white mr-2"></div>
-                                  Eliminando...
-                                </div>
-                              ) : (
-                                'Eliminar'
-                              )}
-                            </button>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-
-                  {quotes.length === 0 ? (
-                    <div className="bg-white rounded-xl p-8 text-center">
-                      <div className="w-16 h-16 flex items-center justify-center bg-gray-100 rounded-full mx-auto mb-4">
-                        <i className="ri-file-list-line text-gray-400 text-2xl"></i>
-                      </div>
-                      <h3 className="text-lg font-medium text-gray-800 mb-2">No hay cotizaciones</h3>
-                      <p className="text-gray-600">Las nuevas cotizaciones aparecerán aquí</p>
-                    </div>
-                  ) : (
-                    quotes.map((quote) => (
-                      <QuoteCard
-                        key={quote.id}
-                        quote={quote}
-                        onStatusUpdate={updateQuoteStatus}
-                        onFinalize={finalizeQuote}
-                      />
-                    ))
-                  )}
-                </div>
+                <QuoteList
+                  quotes={quotes}
+                  showSuccessMessage={showSuccessMessage}
+                  onStatusUpdate={updateQuoteStatus}
+                  onFinalize={finalizeQuote}
+                  onDelete={handleDeleteQuote}
+                />
               )}
 
               {/* Calendar Tab */}
@@ -1092,200 +751,3 @@ export default function DashboardPage() {
   }
 }
 
-// -----------------------------------------------------------------------------
-// QuoteCard component (kept for possible reuse)
-// -----------------------------------------------------------------------------
-function QuoteCard({ quote, onStatusUpdate, onFinalize }: { quote: Quote; onStatusUpdate: (id: string, status: QuoteStatus, estimatedPrice?: number, adminNotes?: string) => void; onFinalize: (quote: Quote) => void }) {
-  const [showDetails, setShowDetails] = useState(false);
-  const [estimatedPrice, setEstimatedPrice] = useState<string>(
-    quote.estimated_price ? String(quote.estimated_price) : ''
-  );
-  const [adminNotes, setAdminNotes] = useState(quote.admin_notes || '');
-
-  const statusColors = {
-    pending: 'bg-yellow-100 text-yellow-800',
-    responded: 'bg-blue-100 text-blue-800',
-    accepted: 'bg-green-100 text-green-800',
-    rejected: 'bg-red-100 text-red-800',
-  };
-
-  const getStatusText = (status: string) => {
-    const statusMap: { [key: string]: string } = {
-      pending: 'Pendiente',
-      responded: 'Respondida',
-      accepted: 'Aceptada',
-      rejected: 'Rechazada',
-    };
-    return statusMap[status] || status;
-  };
-
-  const handleRespond = () => {
-    if (!estimatedPrice) {
-      alert('Por favor ingresa un precio estimado');
-      return;
-    }
-    onStatusUpdate(quote.id, 'responded', parseFloat(estimatedPrice), adminNotes);
-  };
-
-  return (
-    <div className="bg-white rounded-xl shadow-lg p-6 border-l-4 border-purple-400">
-      <div className="flex items-center justify-between mb-4">
-        <div>
-          <h3 className="font-bold text-gray-800 text-lg">{quote.customer_name}</h3>
-          <div className="flex items-center space-x-4 text-sm text-gray-600">
-            <div className="flex items-center">
-              <i className="ri-phone-line mr-1"></i>
-              <span>{quote.customer_phone}</span>
-            </div>
-            {quote.customer_email && (
-              <div className="flex items-center">
-                <i className="ri-mail-line mr-1"></i>
-                <span>{quote.customer_email}</span>
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="text-right">
-          <span className={`px-3 py-1 rounded-full text-xs font-bold ${quoteStatusColors[quote.status as QuoteStatus]} shadow-sm`}>
-            {getStatusText(quote.status)}
-          </span>
-          <p className="text-xs text-gray-500 mt-1">{new Date(quote.created_at).toLocaleDateString('es-ES')}</p>
-        </div>
-      </div>
-
-      <div className="grid grid-cols-2 gap-4 mb-4 text-sm">
-        {quote.occasion && (
-          <div>
-            <span className="text-gray-600">Ocasión:</span>
-            <span className="ml-2 font-medium capitalize">{quote.occasion}</span>
-          </div>
-        )}
-        {quote.servings && (
-          <div>
-            <span className="text-gray-600">Porciones:</span>
-            <span className="ml-2 font-medium">{quote.servings}</span>
-          </div>
-        )}
-        {quote.budget && (
-          <div>
-            <span className="text-gray-600">Presupuesto:</span>
-            <span className="ml-2 font-medium capitalize">{quote.budget}</span>
-          </div>
-        )}
-        {quote.event_date && (
-          <div>
-            <span className="text-gray-600">Fecha del evento:</span>
-            <span className="ml-2 font-medium">{new Date(quote.event_date).toLocaleDateString('es-ES')}</span>
-          </div>
-        )}
-      </div>
-
-      {quote.theme && (
-        <div className="mb-4 p-3 bg-gray-50 rounded-lg">
-          <p className="text-sm font-medium text-gray-700 mb-1">Tema/Estilo:</p>
-          <p className="text-sm text-gray-600">{quote.theme}</p>
-        </div>
-      )}
-
-      {quote.has_reference_photo && quote.photo_description && (
-        <div className="mb-4 p-3 bg-blue-50 rounded-lg border-l-4 border-blue-400">
-          <p className="text-sm font-medium text-blue-800 mb-1">Foto de Referencia:</p>
-          <p className="text-sm text-blue-700">{quote.photo_description}</p>
-          {quote.reference_photo_url && (
-            <div className="mt-2">
-              <img src={quote.reference_photo_url} alt="Referencia" className="max-w-full h-32 object-cover rounded" />
-            </div>
-          )}
-        </div>
-      )}
-
-      {quote.event_details && (
-        <div className="mb-4 p-3 bg-amber-50 rounded-lg border-l-4 border-amber-400">
-          <p className="text-sm font-medium text-amber-800 mb-1">Detalles del Evento:</p>
-          <p className="text-sm text-amber-700">{quote.event_details}</p>
-        </div>
-      )}
-
-      <button onClick={() => setShowDetails(!showDetails)} className="text-purple-600 text-sm font-medium hover:text-purple-700 mb-4">
-        {showDetails ? 'Ocultar opciones' : 'Responder cotización'}
-        <i className={`ri-arrow-${showDetails ? 'up' : 'down'}-s-line ml-1`}></i>
-      </button>
-
-      {showDetails && quote.status === 'pending' && (
-        <div className="border-t pt-4 space-y-4">
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Precio Estimado ($)</label>
-              <input
-                type="number"
-                value={estimatedPrice}
-                onChange={(e) => setEstimatedPrice(e.target.value)}
-                placeholder="85.00"
-                step="0.01"
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 text-sm"
-              />
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-1">Notas Internas</label>
-              <input
-                type="text"
-                value={adminNotes}
-                onChange={(e) => setAdminNotes(e.target.value)}
-                placeholder="Notas adicionales..."
-                className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-purple-500 text-sm"
-              />
-            </div>
-          </div>
-
-          <div className="flex space-x-2">
-            <button
-              onClick={handleRespond}
-              className="flex-1 bg-gradient-to-r from-green-500 to-emerald-500 text-white py-2 px-4 rounded-lg text-sm font-bold hover:from-green-600 hover:to-emerald-600 transition-all"
-            >
-              <i className="ri-check-line mr-2"></i>
-              Enviar Respuesta
-            </button>
-            <button
-              onClick={() => onStatusUpdate(quote.id, 'rejected')}
-              className="flex-1 bg-gradient-to-r from-red-500 to-red-600 text-white py-2 px-4 rounded-lg text-sm font-bold hover:from-red-600 hover:to-red-700 transition-all"
-            >
-              <i className="ri-close-line mr-2"></i>
-              Rechazar
-            </button>
-            <button
-              onClick={() => window.open(`tel:${quote.customer_phone}`, '_self')}
-              className="bg-gradient-to-r from-blue-500 to-blue-600 text-white py-2 px-4 rounded-lg text-sm font-bold hover:from-blue-600 hover:to-blue-700 transition-all"
-            >
-              <i className="ri-phone-line"></i>
-            </button>
-          </div>
-        </div>
-      )}
-
-      {quote.estimated_price && (
-        <div className="mt-4 p-3 bg-green-50 rounded-lg border-l-4 border-green-400">
-          <div className="flex justify-between items-center">
-            <div>
-              <p className="text-sm font-medium text-green-800">Precio Cotizado:</p>
-              <p className="text-lg font-bold text-green-700">${quote.estimated_price}</p>
-            </div>
-            {quote.admin_notes && (
-              <div className="text-right">
-                <p className="text-xs text-green-600">Notas: {quote.admin_notes}</p>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-
-      {quote.status === 'responded' && (
-        <button
-          onClick={() => onFinalize(quote)}
-          className="mt-4 w-full bg-gradient-to-r from-pink-500 to-purple-500 text-white py-2 px-4 rounded-lg text-sm font-bold hover:from-pink-600 hover:to-purple-600 transition-all"
-        >
-          Finalizar y Crear Orden
-        </button>
-      )}
-    </div>
-  );
-}
